Moved article URL functionality into its own class, no more export through article service. This one remains, though empty at the moment. It will, at some later point, load and store articles from/to the persistence thing.
