--- src-app/services/article/ArticleService.live.ts
@@ -1,3 +1,0 @@
-export { default as parseURL } from './parseURL';
-
-console.log('ArticleService.live');
