--- src-app/services/article/ArticleService.mock.ts
@@ -1,3 +1,0 @@
-export { default as parseURL } from './parseURL';
-
-console.log('ArticleService.mock');
