+++ src-app/services/article/ArticleService.ts
@@ -1,7 +1,5 @@
 interface ArticleService {
 
-	parseURL(rawURL : string) : string;
-
 }
 
 export default ArticleService;
